itunes dll for windows now playing now supports StreamTitles (Fixes issue 1719) FreeBSD 6 and 7 can now download and install a JRE. (Fixes issue 1722) Launcher and Installer now source the ~/.profile file when looking for java. Correct some java error messages in the installer
(All backported to 0.6 branch aswell)


git-svn-id: http://svn.dmdirc.com/branches/0.6@4620 00569f92-eb28-0410-84fd-f71c24880f43

+++ installer/linux/installjre.sh
@@ -51,6 +51,8 @@ if [ "${PIDOF}" = "" ]; then
 	fi;
 fi;
 
+ISFREEBSD=`uname -s | grep -i FreeBSD`
+
 ## Helper Functions
 if [ "${PIDOF}" != "" ]; then
 	ISKDE=`${PIDOF} -x -s kdeinit`
@@ -124,17 +126,34 @@ questiondialog() {
 
 showLicense() {
 	# Get License Text
-	FILE=`mktemp -p ${PWD} license.XXXXXXXXXXXXXX`
-	
-	# Location of license start
-	STARTLINE=`grep ${GREPOPTS} "^more <<\"EOF\"$" jre.bin`
-	STARTLINE=$((${STARTLINE%%:*} + 1))
-	# Location of license end
-	ENDLINE=`grep ${GREPOPTS} "Do you agree to the above license terms?" jre.bin`
-	ENDLINE=$((${ENDLINE%%:*} - 2))
-	
-	head -n ${ENDLINE} jre.bin | tail ${TAILOPTS}${STARTLINE} > ${FILE}
-	
+	FILE=`mktemp license.XXXXXXXXXXXXXX`
+	if [ "${ISFREEBSD}" != "" ]; then
+		WGET=`which wget`
+		FETCH=`which fetch`
+		CURL=`which curl`
+		
+		ARCH=`uname -m`
+		RELEASE=`uname -r`
+		URL="http://www.dmdirc.com/getjavalicense/`uname -s`/${ARCH}/${RELEASE}"
+		
+		if [ "${WGET}" != "" ]; then
+			${WGET} -q -O ${FILE} ${URL}
+		elif [ "${FETCH}" != "" ]; then
+			${FETCH} -q -o ${FILE} ${URL}
+		elif [ "${CURL}" != "" ]; then
+			${CURL} -s -o ${FILE} ${URL}
+		fi;
+	else
+		# Location of license start
+		STARTLINE=`grep ${GREPOPTS} "^more <<\"EOF\"$" jre.bin`
+		STARTLINE=$((${STARTLINE%%:*} + 1))
+		# Location of license end
+		ENDLINE=`grep ${GREPOPTS} "Do you agree to the above license terms?" jre.bin`
+		ENDLINE=$((${ENDLINE%%:*} - 2))
+		
+		head -n ${ENDLINE} jre.bin | tail ${TAILOPTS}${STARTLINE} > ${FILE}
+	fi;
+
 	# Send text to console.
 	echo ""
 	echo "-----------------------------------------------------------------------"
@@ -169,9 +188,13 @@ messagedialog "Java Install" "Before java can be installed, please review the fo
 showLicense
 questiondialog "Java Install" "Do you agree to the Java License?"
 if [ $? -eq 0 ]; then
-	# Look to see where the JRE wants to install to
-	JREJAVAHOME=`grep ${GREPOPTS} "^javahome=" jre.bin`
- 	JREJAVAHOME=${JREJAVAHOME##*=}
+	if [ "${ISFREEBSD}" != "" ]; then
+		JREJAVAHOME="diablo-jre1.6.0_07"
+	else
+		# Look to see where the JRE wants to install to
+		JREJAVAHOME=`grep ${GREPOPTS} "^javahome=" jre.bin`
+		JREJAVAHOME=${JREJAVAHOME##*=}
+	fi;
 	
 	echo "JREJAVAHOME: ${JREJAVAHOME}"
 	
@@ -188,33 +211,40 @@ if [ $? -eq 0 ]; then
 	echo "installdir: ${installdir}"
 	
 	if [ ! -e ${installdir} ]; then
-		# Hack jre.bin to allow us to install without asking for a license, or failing
-		# the checksum.
-		
-		# Location of license start
-		STARTLINE=`grep ${GREPOPTS} "^more <<\"EOF\"$" jre.bin`
-		STARTLINE=${STARTLINE%%:*}
-		# Location of license end
-		ENDLINE=`grep ${GREPOPTS} "If you don't agree to the license you can't install this software" jre.bin`
-		ENDLINE=$((${ENDLINE%%:*} + 3))
-		# Location of checksum start
-		CSSTARTLINE=`grep ${GREPOPTS} "^if \[ -x /usr/bin/sum \]; then$" jre.bin`
-		CSSTARTLINE=${CSSTARTLINE%%:*}
-		# Location of checksum end
-		CSENDLINE=`grep ${GREPOPTS} "Can't find /usr/bin/sum to do checksum" jre.bin`
-		CSENDLINE=$((${CSENDLINE%%:*} + 2))
-		# Location of script end
-		SCENDLINE=`grep ${GREPOPTS} "^echo \"Done.\"$" jre.bin`
-		SCENDLINE=$((${SCENDLINE%%:*} + 2 - (${ENDLINE} - ${STARTLINE}) - (${CSENDLINE} - ${CSSTARTLINE})))
-		# Remove the license and checksum stuff!
-		head -n $((${STARTLINE} -1)) jre.bin > jre.bin.tmp
-		tail ${TAILOPTS}$((${ENDLINE})) jre.bin | head -n $((${CSSTARTLINE} -1 - ${ENDLINE})) >> jre.bin.tmp
-		echo "tail \${tail_args} +${SCENDLINE} \"\$0\" > \$outname" >> jre.bin.tmp
-		tail ${TAILOPTS}$((${CSENDLINE})) jre.bin >> jre.bin.tmp
-		
 		messagedialog "Java Install" "Java install will begin when you press OK.\nThis may take some time, so please wait.\n\nYou will be informed when the installation is completed."
-		yes | sh jre.bin.tmp
-		rm -Rf jre.bin.tmp
+		if [ "${ISFREEBSD}" != "" ]; then
+			mv jre.bin jre.tar.bz2
+			tar -jxvf jre.tar.bz2
+			mv jre.tar.bz2 jre.bin
+		else
+			# Hack jre.bin to allow us to install without asking for a license, or failing
+			# the checksum.
+			
+			# Location of license start
+			STARTLINE=`grep ${GREPOPTS} "^more <<\"EOF\"$" jre.bin`
+			STARTLINE=${STARTLINE%%:*}
+			# Location of license end
+			ENDLINE=`grep ${GREPOPTS} "If you don't agree to the license you can't install this software" jre.bin`
+			ENDLINE=$((${ENDLINE%%:*} + 3))
+			# Location of checksum start
+			CSSTARTLINE=`grep ${GREPOPTS} "^if \[ -x /usr/bin/sum \]; then$" jre.bin`
+			CSSTARTLINE=${CSSTARTLINE%%:*}
+			# Location of checksum end
+			CSENDLINE=`grep ${GREPOPTS} "Can't find /usr/bin/sum to do checksum" jre.bin`
+			CSENDLINE=$((${CSENDLINE%%:*} + 2))
+			# Location of script end
+			SCENDLINE=`grep ${GREPOPTS} "^echo \"Done.\"$" jre.bin`
+			SCENDLINE=$((${SCENDLINE%%:*} + 2 - (${ENDLINE} - ${STARTLINE}) - (${CSENDLINE} - ${CSSTARTLINE})))
+			# Remove the license and checksum stuff!
+			head -n $((${STARTLINE} -1)) jre.bin > jre.bin.tmp
+			tail ${TAILOPTS}$((${ENDLINE})) jre.bin | head -n $((${CSSTARTLINE} -1 - ${ENDLINE})) >> jre.bin.tmp
+			echo "tail \${tail_args} +${SCENDLINE} \"\$0\" > \$outname" >> jre.bin.tmp
+			tail ${TAILOPTS}$((${CSENDLINE})) jre.bin >> jre.bin.tmp
+		
+			yes | sh jre.bin.tmp
+			rm -Rf jre.bin.tmp
+		fi;
+		
 		mv ${JREJAVAHOME} ${installdir}
 		
 		if [ "0" = "${UID}" ]; then
@@ -230,13 +260,12 @@ if [ $? -eq 0 ]; then
 			ln -s ${installdir}/bin/java ${PWD}/java-bin
 		else
 			# Add to path.
-			if [ -e ${HOME}/.profile ]; then
-				echo "" >> ${HOME}/.profile
-				echo "# set PATH so it includes user's private java if it exists" >> ${HOME}/.profile
-				echo "if [ -d ~/${JREJAVAHOME}/bin ] ; then" >> ${HOME}/.profile
-				echo "	PATH=~/${JREJAVAHOME}/bin:\${PATH}" >> ${HOME}/.profile
-				echo "fi" >> ${HOME}/.profile
-			fi
+			echo "" >> ${HOME}/.profile
+			echo "# set PATH so it includes user's private java if it exists" >> ${HOME}/.profile
+			echo "if [ -d ~/${JREJAVAHOME}/bin ] ; then" >> ${HOME}/.profile
+			echo "	PATH=~/${JREJAVAHOME}/bin:\${PATH}" >> ${HOME}/.profile
+			echo "fi" >> ${HOME}/.profile
+			
 			if [ -e ${HOME}/.cshrc ]; then
 				echo "" >> ${HOME}/.cshrc
 				echo "# set PATH so it includes user's private java if it exists" >> ${HOME}/.cshrc
